added mapper, reducer comparator tasks

+++ mapreduce/src/main/java/com/css534/parallel/Vector2f.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Vector2f {
 
-    private int xx;
-    private int yy;
+    private double xx;
+    private double yy;
 
 }
